chore: update lint commands
