+++ package.json
@@ -21,8 +21,8 @@
     "affected": "nx affected",
     "affected:apps": "ts-node --project tools/tsconfig.json tools/nx/get-projects-cli.ts -t app -p",
     "affected:libs": "ts-node --project tools/tsconfig.json tools/nx/get-projects-cli.ts -t lib -p",
-    "affected:lint:old": "nx affected:lint",
-    "affected:lint": "node tools/nx/fast-lint.js -v",
+    "affected:lint": "nx affected:lint",
+    "affected:lint:fast": "ts-node --project tools/tsconfig.json tools/nx/fast-lint.js -v",
     "affected:dep-graph": "nx affected:dep-graph",
     "affected:build:docker": "nx affected --target=docker-build",
     "affected:e2e": "nx affected:e2e",
